+++ tests/TestCase/Core/GlobalFunctionsTest.php
@@ -32,6 +32,26 @@ class GlobalFunctionsTest extends TestCase
     }
 
     /**
+     * Test for `isWin()` global function on Unix
+     * @group onlyUnix
+     * @test
+     */
+    public function testIsWinOnUnix()
+    {
+        $this->assertFalse(isWin());
+    }
+
+    /**
+     * Test for `isWin()` global function on Windows
+     * @group onlyWindows
+     * @test
+     */
+    public function testIsWinOnWin()
+    {
+        $this->assertTrue(isWin());
+    }
+
+    /**
      * Test for `rtr()` global function
      * @test
      */
@@ -51,14 +71,24 @@ class GlobalFunctionsTest extends TestCase
     }
 
     /**
-     * Test for `which()` global function
+     * Test for `which()` global function on Unix
+     * @group onlyUnix
      * @test
      */
-    public function testWhich()
+    public function testWhichOnUnix()
     {
-        $expected = exec('which cat');
-        $this->assertEquals($expected, which('cat'));
+        $this->assertEquals('/bin/cat', which('cat'));
+        $this->assertNull(which('noExistingBin'));
+    }
 
+    /**
+     * Test for `which()` global function on Windows
+     * @group onlyWindows
+     * @test
+     */
+    public function testWhichOnWindws()
+    {
+        $this->assertEquals('"C:\Program Files\Git\usr\bin\cat.exe"', which('cat'));
         $this->assertNull(which('noExistingBin'));
     }
 }