+++ tests/bootstrap.php
@@ -90,9 +90,16 @@ Cache::setConfig([
     ],
 ]);
 
-ConnectionManager::setConfig('test', ['url' => 'mysql://root@localhost/test']);
+if (!getenv('db_dsn')) {
+    putenv('db_dsn=mysql://root:@localhost/test');
+}
+if (!getenv('db_dsn_postgres')) {
+    putenv('db_dsn_postgres=postgres://postgres@localhost/travis_ci_test');
+}
+
+ConnectionManager::setConfig('test', ['url' => getenv('db_dsn')]);
+ConnectionManager::setConfig('test_postgres', ['url' => getenv('db_dsn_postgres')]);
 ConnectionManager::setConfig('test_sqlite', ['url' => 'sqlite:///' . TMP . 'example.sq3']);
-ConnectionManager::setConfig('test_postgres', ['url' => 'postgres://postgres@localhost/travis_ci_test']);
 
 Configure::write('DatabaseBackup.connection', 'test');
 Configure::write('DatabaseBackup.target', TMP . 'backups');